+++ GFF2026A/Assets/Scripts/Firebase/Auth.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 95f6ec1372acb4b46af25ab4c1632792
